use cookies to remember view mode
