+++ actions/index.ts
@@ -1,3 +1,4 @@
+export * from "@/actions/cookies";
 export * from "@/actions/create";
 export * from "@/actions/email";
 export * from "@/actions/get";
