+++ types/types.d.ts
@@ -56,3 +56,5 @@ export interface DropdownOption {
   label: string;
   value: string;
 }
+
+export type ViewMode = "private" | "global";
